backend/requirements.txt
click>=8.0.0
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.29.0
psycopg[binary]>=3.2.0
alembic>=1.13.0
pydantic-settings>=2.6.0
bcrypt>=4.0.1
PyJWT>=2.9.0
greenlet>=3.1.0
pgvector>=0.3.6
email-validator>=2.2.0
cryptography>=42.0.0
httpx>=0.27.0
structlog>=24.0.0
slowapi>=0.1.9
pycrdt>=0.12.0
pycrdt-websocket>=0.16.0
minio>=7.2.0
pypdf>=5.0.0
llama-index-core==0.14.21
litellm>=1.55.0,<2
